+++ src/index.html
@@ -164,8 +164,9 @@
             <div id="Loading_Text">Loading</div>
         </div>
     </div>
-    <script src="./scripts/com.js"></script>
     <script src="./scripts/app.js"></script>
+    <script src="./scripts/serverselect.js"></script>
+    <script src="./scripts/createserver.js"></script>
 </body>
 
 </html>
