+++ src/scripts/serverselect.js
@@ -1,0 +1,33 @@
+const updateServerList = (dir) => {
+    let subdirs = fs.readdirSync(dir, { withFileTypes: true, })
+        .filter(dirent => dirent.isDirectory())
+        .map(dirent => dirent.name);
+
+    document.querySelector("#ServerSelect_Label").innerText = `${subdirs.length} Server${subdirs.length === 1 ? "" : "s"} Found`;
+
+    let layout = new Array(subdirs.length).fill(null);
+    for (let i = 0; i < layout.length; i++) layout[i] = {
+        tag: "div",
+        children: [
+            {
+                tag: "img",
+                src: fs.existsSync(path.join(DIR.SERVERS, subdirs[i], "./server-icon.png")) ? path.join(DIR.SERVERS, subdirs[i], "./server-icon.png") : "./default.png",
+                style: { verticalAlign: "middle", },
+                draggable: false,
+            },
+            {
+                tag: "span",
+                content: subdirs[i],
+                style: { marginLeft: "15px", },
+            }
+        ],
+    };
+
+    while (document.querySelector("#ServerSelect_List").children.length > 0) document.querySelector("#ServerSelect_List").children[0].remove();
+    createLayout(layout, document.querySelector("#ServerSelect_List"));
+};
+fs.watch(DIR.SERVERS, () => updateServerList(DIR.SERVERS));
+updateServerList(DIR.SERVERS);
+
+document.querySelector("#ServerSelect_CreateServer").onclick = () => showPanel("CreateServer");
+document.querySelector("#ServerSelect_OpenServerFolder").onclick = () => openExternal(DIR.SERVERS);
